Refactors alert types and context management by consolidating type definitions and removing unwanted index.d.ts file
