+++ src/index.ts
@@ -1,2 +1,3 @@
 export {default as AlertProvider} from "./AlertProvider";
 export {default as useAlert} from "./useAlert";
+export type { Alert, AlertType, AlertProviderProps } from "./types";
